jsteg: fix issue with jstegdev_init
There was problem with initialization. User wasn't able to mount
after addition some pictures to fs-directory/

+++ src/include/bdev_jsteg/jstegentry.h
@@ -54,7 +54,7 @@ jdecompress_destroy(jdecompress_state* cinfo_ptr);
 
 
 typedef struct jsteg_entry_t {
-		uint16_t jindex;
+		int32_t jindex;
 		FILE* file;
 		
 		size_t start;
